# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Quin Travel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Register a user",
                "description": "Returns new user",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "description": "Your full name",
                                        "type": "string",
                                        "example": "Nguyen van a"
                                    },
                                    "email": {
                                        "description": "Your email address",
                                        "type": "string",
                                        "example": "test@gmail.com"
                                    },
                                    "password": {
                                        "description": "Your password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "confirm password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/auth/update_profile": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update your profile information",
                "description": "Returns new profile information",
                "operationId": "update_profile",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "description": "Your full name",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "Your address",
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "description": "Your phone number ",
                                        "type": "integer"
                                    },
                                    "avatar": {
                                        "description": "Your avatar ",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/auth/change_password": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Change password for current",
                "description": "Returns status of the change password",
                "operationId": "change_password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "description": "Your current password",
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "description": "Your new password",
                                        "type": "string"
                                    },
                                    "new_password_confirmation": {
                                        "description": "Your new_password_confirmation ",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Login",
                "description": "Returns user",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Your email address",
                                        "type": "string",
                                        "example": "test@gmail.com"
                                    },
                                    "password": {
                                        "description": "Your password",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get User Information",
                "description": "Returns user information",
                "operationId": "getme",
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Logout account",
                "description": "Returns status",
                "operationId": "logout",
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Refresh token user",
                "description": "Returns status",
                "operationId": "refresh",
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "JWT Token (Bearer <token>)"
            }
        }
    }
}